Adds option to use doclava doclet instead of standard as a base one
Although, doclava has not been updated for a while
it still works and has nice features like:
 * embedded search “from the box”
 * direct links to classes
 * extra keywords like `@hide`

Since markdown-doclet doesn’t actually replace the standard one,
but works on top of it, it’s quite easy to substitute it with doclava.

Current implementation checks if doclava doclet is in classpath and uses it if it is.
If doclava is not in classpath, standard doclet is used.

Example of usage in `build.gradle`:
```
        configurations {
            doclava
        }

        dependencies {
            doclava 'ch.raffael.mddoclet.markdown-doclet:jdk8:2.0+'
            doclava 'net.sourceforge.plantuml:plantuml:8059'
            doclava 'com.google.doclava:doclava:1.0.6'
        }

        javadoc {
            dependsOn project.configurations.doclava
            title = null
            options.doclet = "ch.raffael.mddoclet.MarkdownDoclet"
            options.docletpath = configurations.doclava.files.asType(List)
            options.addStringOption 'hdf project.name', “My Project“
        }

```